gitagyaan/requirements.txt
google-generativeai
faiss-cpu
streamlit
numpy
pandas
tqdm
